Now posts and comments are not cached by default

+++ app/src/main/java/com/elfen/letsred/models/Comment.kt
@@ -1,18 +1,18 @@
 package com.elfen.letsred.models
 
-sealed class Comment(open val order: Float, open val depth: Int) {
+sealed class Comment(open val order: Int, open val depth: Int) {
     data class Content(
         val id: String,
         val body: String,
         val authorId: String?,
         val authorUsername: String?,
         override val depth: Int,
-        override val order: Float,
+        override val order: Int,
     ): Comment(order, depth)
 
     data class More(
         val id: String,
         override val depth: Int,
-        override val order: Float
+        override val order: Int
     ): Comment(order, depth)
 }
